requirements.txt
appier==1.18.4
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
google-api==0.1.12
google-api-core==1.7.0
google-auth==1.6.2
googleapis-common-protos==1.5.6
grpcio==1.17.1
grpcio-tools==1.17.1
idna==2.8
ijson==2.3
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
tqdm==4.29.1
urllib3==1.24.1
